style: what clippy says, across the tree
Ninety-nine findings, all of them answered. The mechanical ones: a
`clone` on a `Copy` register, a redundant field name outside a lowered
unit, `% n == 0` as `is_multiple_of`, an unused import or variable, a
`mut` nothing writes, a value assigned and never read, a manual range
test, `map_or(false, ..)` as `is_some_and`, a range loop as an
`iter_mut().enumerate()`, `last()` on a double-ended iterator as
`next_back()`.

Four are allowed where they fire, each with its reason beside it,
because the lowering forces the form: a struct literal's field is
`name: value` and not a shorthand (issue 235), a `select!` arm's
alternatives are compared one by one and not as a range, a lowered
function's parameters are its wires and there are as many as the
hardware has, and `U::mul::<M>` is not `std::ops::Mul`, which takes no
width. Two more are allowed for what they are: the header that takes
the widths of the rows below it, and a `drop` that ends a closure's
borrow.

An `allow` goes above `#[lower]` and never between it and the `fn`,
which hides the function from the scan that finds it (issue 234).

Assistant: the change was written by Claude Opus 5 under the author's
direction.

Prompts: Go and continue looping without approval

+++ lib/parts/src/bus/axi.rs
@@ -318,6 +318,10 @@ pub struct AxiHost<
     pub busy: Reg<U<NIDS>>,
 }
 
+// A struct literal's field is `name: value` to the lowering,
+// which reads no shorthand (issue 235), so `id: id` stays as it is
+// written.
+#[allow(clippy::redundant_field_names)]
 #[lower]
 impl<
         const A: usize,
@@ -1808,7 +1812,7 @@ mod tests {
                 x ^= x >> 17;
                 x ^= x << 5;
                 let words = (x as usize % 4) + 1;
-                let addr = ((x >> 8) % 8) as u32 * 4;
+                let addr = ((x >> 8) % 8) * 4;
                 if x & 0x40 == 0 {
                     let vs: Vec<U<32>> =
                         (0..words).map(|i| U::from(next + i as u32)).collect();
